Default mock PCS config
